fix(office-preview): hoist ZIP magic check, DocxViewer now guards too
Move ZIP_MAGIC + isZipBuffer to src/shared/zip-magic.ts and import from both
viewers. .docx and .xlsx are both ZIP containers, so a truncated/non-ZIP
payload should fail with a clear 'not a zip archive' error from both, not
surface as the viewer-specific 'corrupt or encrypted' copy in DocxViewer.
